+++ app/src/main/java/org/linphone/ui/main/chat/viewmodel/ConversationDocumentsListViewModel.kt
@@ -79,8 +79,17 @@ class ConversationDocumentsListViewModel
             val size = documentContent.size.toLong()
             val timestamp = documentContent.creationTimestamp
             if (path.isNotEmpty() && name.isNotEmpty()) {
-                // TODO FIXME: we don't have the ephemeral info at Content level, using the chatRoom info even if content ephemeral status may or may not be different...
-                val ephemeral = chatRoom.isEphemeralEnabled
+                val messageId = documentContent.relatedChatMessageId
+                val ephemeral = if (messageId != null) {
+                    val chatMessage = chatRoom.findMessage(messageId)
+                    if (chatMessage == null) {
+                        Log.w("$TAG Failed to find message using ID [$messageId] related to this content, can't get real info about being related to ephemeral message")
+                    }
+                    chatMessage?.isEphemeral ?: chatRoom.isEphemeralEnabled
+                } else {
+                    Log.e("$TAG No chat message ID related to this content, can't get real info about being related to ephemeral message")
+                    chatRoom.isEphemeralEnabled
+                }
 
                 val model =
                     FileModel(path, name, size, timestamp, isEncrypted, originalPath, ephemeral) {